Refactor shared app package into common module
