+++ common/admin/views/dashboard.py
