+++ common/text_extract.py
