+++ services/matching/Dockerfile
@@ -16,7 +16,7 @@ RUN --mount=type=cache,target=/root/.cache/pip \
 RUN --mount=type=cache,target=/root/.cache/pip \
     pip install --prefer-binary -r requirements.txt
 
-COPY app ./app
+COPY common ./common
 COPY templates ./templates
 COPY services ./services
